requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
statsbombpy==1.4.2
pandas==2.1.3
supabase==2.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
